stackfusion/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
python-dotenv==0.13.0
django-phone-field==1.8.0
